requirements.txt
amqp==2.4.2
appdirs==1.4.4
asgiref==3.7.2
attrs==21.2.0
Babel==2.9.1
billiard==3.5.0.5
business-calendar==0.2.1
cached-property==1.5.1
celery==4.2.1
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==1.11.29
django-autocomplete-light==3.5.1
django-celery-results==1.0.4
django-ckeditor==5.2.2
django-cookies-samesite==0.9.0
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
django-nested-admin==3.2.4
django-nested-inline==0.4.6
django-rest-swagger==2.1.2
django-salesforce==3.1
django-stdimage==3.0.0
django-suit==0.2.26
djangorestframework==3.7.3
et-xmlfile==1.0.1
factory-boy==3.3.0
Faker==19.2.0
flower==0.9.2
gunicorn==20.0.4
humanize==0.5.1
idna==2.6
isodate==0.6.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10
kombu==4.4.0
lxml==4.6.3
MarkupSafe==1.1.1
numpy==1.25.2
olefile==0.44
openapi-codec==1.3.2
openpyxl==3.0.3
pandas==2.0.3
Pillow==8.2.0
polyfactory==2.6.3
progressbar2==3.34.3
psycopg2-binary==2.8.6
pyactiveresource==2.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
pydantic==1.10.11
PyJWT==2.4.0
pyodbc==4.0.26
pyStrich==0.8
python-dateutil==2.8.2
python-dotenv==0.14.0
python-ldap==3.2.0
python-magic==0.4.15
python-memcached==1.59
python-monkey-business==1.0.0
python-utils==2.2.0
pytz==2023.3
PyYAML==3.12
querystring-parser==1.2.3
redis==3.2.0
requests==2.18.4
requests-toolbelt==0.8.0
setproctitle==1.1.10
ShopifyAPI==3.1.0
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
tornado==5.1.1
typing_extensions==4.5.0
tzdata==2023.3
ua-parser==0.10.0
uritemplate==3.0.0
urllib3==1.22
vine==1.2.0
xlrd==1.2.0
xlwt==1.3.0
yolk3k==0.9
zeep==3.1.0
